refactor: setup loader api
