+++ cognite/neat/_client/_api/data_modeling_loaders.py
@@ -1,6 +1,7 @@
+from abc import ABC, abstractmethod
 from collections.abc import Callable, Sequence
 from graphlib import TopologicalSorter
-from typing import Any, Literal, cast
+from typing import TYPE_CHECKING, Any, Generic, Literal, TypeVar, cast
 
 from cognite.client import CogniteClient
 from cognite.client.data_classes import filters
@@ -8,6 +9,7 @@ from cognite.client.data_classes._base import (
     T_CogniteResourceList,
     T_WritableCogniteResource,
     T_WriteClass,
+    WriteableCogniteResourceList,
 )
 from cognite.client.data_classes.data_modeling import (
     Container,
@@ -45,7 +47,50 @@ from cognite.client.data_classes.data_modeling.views import (
 from cognite.client.exceptions import CogniteAPIError
 from cognite.client.utils.useful_types import SequenceNotStr
 
-from .loader_base import T_ID, ResourceLoader, T_WritableCogniteResourceList
+from cognite.neat._shared import T_ID
+
+if TYPE_CHECKING:
+    from cognite.neat._client.api_client import NeatClient
+
+T_WritableCogniteResourceList = TypeVar("T_WritableCogniteResourceList", bound=WriteableCogniteResourceList)
+
+
+class ResourceLoader(
+    ABC,
+    Generic[T_ID, T_WriteClass, T_WritableCogniteResource, T_CogniteResourceList, T_WritableCogniteResourceList],
+):
+    resource_name: str
+
+    def __init__(self, client: CogniteClient) -> None:
+        self.client = client
+
+    @classmethod
+    @abstractmethod
+    def get_id(cls, item: T_WriteClass | T_WritableCogniteResource | dict | T_ID) -> T_ID:
+        raise NotImplementedError
+
+    @classmethod
+    def get_ids(cls, items: Sequence[T_WriteClass | T_WritableCogniteResource]) -> list[T_ID]:
+        return [cls.get_id(item) for item in items]
+
+    @abstractmethod
+    def create(self, items: Sequence[T_WriteClass]) -> T_WritableCogniteResourceList:
+        raise NotImplementedError
+
+    @abstractmethod
+    def retrieve(self, ids: SequenceNotStr[T_ID]) -> T_WritableCogniteResourceList:
+        raise NotImplementedError
+
+    @abstractmethod
+    def update(self, items: Sequence[T_WriteClass]) -> T_WritableCogniteResourceList:
+        raise NotImplementedError
+
+    @abstractmethod
+    def delete(self, ids: SequenceNotStr[T_ID]) -> list[T_ID]:
+        raise NotImplementedError
+
+    def are_equal(self, local: T_WriteClass, remote: T_WritableCogniteResource) -> bool:
+        return local == remote.as_write()
 
 
 class DataModelingLoader(
@@ -376,3 +421,12 @@ class DataModelLoader(DataModelingLoader[DataModelId, DataModelApply, DataModel,
         )
 
         return local_dumped == cdf_resource_dumped
+
+
+class DataModelLoaderAPI:
+    def __init__(self, client: NeatClient) -> None:
+        self._client = client
+        self.spaces = SpaceLoader(client)
+        self.views = ViewLoader(client)
+        self.containers = ContainerLoader(client)
+        self.data_models = DataModelLoader(client)
