+++ cognite/neat/_workflows/steps/lib/current/rules_validator.py
@@ -5,7 +5,7 @@ from typing import ClassVar
 
 from cognite.client import CogniteClient
 
-from cognite.neat._client.api.data_modeling_loaders import ViewLoader
+from cognite.neat._client._api.data_modeling_loaders import ViewLoader
 from cognite.neat._issues import NeatIssueList
 from cognite.neat._issues.errors import ResourceNotFoundError, WorkflowStepNotInitializedError
 from cognite.neat._issues.formatters import FORMATTER_BY_NAME
